Add check for systemd-resolved and remove epel-release.
The Centos OS version that I tested this script on did not have
systemd-resolved running so no issues cropped up at the time.

I wanted to test the install scripts on another yum/rpm style distro so
I tested the script on Fedora.

It failed as there was no check to see if systemd-resolved was running
and to disable it. This prevented Unbound from being able to bind a
socket on port 53.

I also noticed that on Fedora, at least for recent versions installing
the epel-release package is a bit more involved compared to say Centos
7.

I then discovered that the epel-release repo/packages are not actually
needed at all to install the dependencies I need for Unbound in recent
versions of Centos and Fedora.

Considering that this script builds Unbound from source which is as
close to bleeding edge you can get.

I am making the trade off that I will assume the dependencies I want to
install can be found in the default distro repos after an update call is
performed.

+++ install-apt-get.sh
@@ -35,8 +35,10 @@ chmod 755 /etc/systemd/system/unbound.service | tee -a run.log
 wget 'https://raw.githubusercontent.com/buggysolid/unbound-config/main/resolv.conf' -O "$HOME/resolv.conf" | tee -a run.log
 mv "$HOME/resolv.conf" /etc/resolv.conf | tee -a run.log
 chmod 744 /etc/resolv.conf | tee -a run.log
-systemctl stop systemd-resolved | tee -a run.log
-systemctl disable systemd-resolved | tee -a run.log
+if [[ $(systemctl is-enabled systemd-resolved) == 'enabled' ]]; then
+    systemctl stop systemd-resolved | tee -a run.log
+    systemctl disable systemd-resolved | tee -a run.log
+fi
 systemctl enable unbound | tee -a run.log
 systemctl restart unbound | tee -a run.log
 systemctl status unbound | tee -a run.log
